fix(extractor): an enum-ordinal overflow hid NINE units from every index
TryWalkEnum kept the running enum ordinal in an Integer. The counter holds
"last ordinal + 1", whose range is Low(Integer)..High(Integer)+1 -- one value
wider than an Integer, and the `= $7FFFFFFF` FORCE_DWORD idiom (which pins a
Delphi enum to 32 bits) hits exactly that value. The member itself was emitted
correctly; the unconditional Inc on the next line then raised EIntOverflow
under the Debug build's DCC_IntegerOverflowCheck.

That threw out of Parser.Parse BEFORE Core.Indexer's OpenFileTx, so the FILE
never got a row -- no symbols, no refs, no literals, no docs. Both catch sites
print `SKIP <path>: EIntOverflow` , increment no error counter and leave the
exit code at 0, and because the file has no row it is never up-to-date, so
every later run re-attempted and re-skipped it.

Absent from every index since 18bbbc8 (2026-07-26): Winapi.D3D10,
Winapi.D3DCommon, Winapi.D3DCompiler, Winapi.D3DX10, Winapi.DXGI1_2,
WinAPI.Media, WinAPI.UI.Core and dxFontFile (x2 platforms).

Fix is one type: Integer -> Int64 on the counter. Int64 prints identical digits
for every value in Integer range, so every file that indexes today is
byte-identical after; only the nine gain content.

DRAGLINT_EXTRACTOR_VERSION 1.13.0-alpha -> 1.14.0-alpha. The parser now
produces a full parse where it produced nothing, which is the largest possible
change in what extraction yields, so this rides a bump and a full re-parse. The
cheaper hash-only path was considered and rejected in the baseline entry rather
than left unsaid.

Guarded by tests\autotest\run_enum_ordinal_overflow.ps1, MEASURED RED then
GREEN with the same runner against two engines: 8 PASS / 8 FAIL against a
pre-fix copy taken before any edit, 16 PASS / 0 FAIL after. Its four controls
(ctA=0, ctB=128, ctC=129, and a symbol that does not exist) pass on BOTH, so
the runs differ in the engine and nothing else. Also A/B'd on the real
Winapi.D3D10.pas: 1 skip / 0 hits before, 0 skips / 1 hit after.

Full battery 497 pass / 0 fail / 0 timeout.

+++ src/parser/DRagLint.Parser.Delphi13.pas
@@ -967,7 +967,18 @@ begin
   // enum semantics. A non-integer initializer (rare: `= SomeConst`) leaves the
   // running counter (best-effort; absence-over-wrong is not possible for a bare
   // number, so we keep the positional value).
-  var EnumOrd: Integer:= 0;
+  // Int64, NOT Integer, and this is load-bearing. The counter holds "last ordinal
+  // + 1", whose range is Low(Integer)..High(Integer)+1 -- one value wider than an
+  // Integer. The `= $7FFFFFFF` FORCE_DWORD idiom (which pins an enum to 32 bits)
+  // hits exactly that value, and the Inc below then raised EIntOverflow under the
+  // Debug build's DCC_IntegerOverflowCheck. That threw out of Parse BEFORE
+  // OpenFileTx, so the FILE got no row at all: nine units (Winapi.D3D10,
+  // D3DCommon, D3DCompiler, D3DX10, DXGI1_2, WinAPI.Media, WinAPI.UI.Core,
+  // dxFontFile x2) were absent from every index from 2026-07-26 to 2026-09-09,
+  // with only a SKIP line and exit code 0 to show for it.
+  // THE GENERAL RULE: a value read from SOURCE TEXT must not be incremented in a
+  // type it can saturate. Guarded by tests\autotest\run_enum_ordinal_overflow.ps1.
+  var EnumOrd: Int64:= 0;
   for i:= 0 to EnumNode.NamedChildCount - 1 do
   begin
     ValNode:= EnumNode.NamedChild(i);
